requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
pydantic==2.9.2
apscheduler==3.10.4
requests==2.32.3
jsonpath-ng==1.6.1
python-dotenv==1.0.1
streamlit==1.37.1
pyyaml
